backend/requirements.txt
django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
python-docx>=1.1,<2.0
pypdf>=5.0,<6.0
openai>=1.65.0,<2.0
requests>=2.32.0,<3.0
